requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0

# Text Encoder
transformers>=4.35.0
ftfy>=6.1.1
regex>=2023.10.3

# Data
pycocotools>=2.0.7
albumentations>=1.3.1
Pillow>=10.0.0

# Inference Acceleration
onnx>=1.15.0
onnxruntime>=1.16.0

# Experiment Tracking
tensorboard>=2.15.0
# wandb>=0.16.0  # optional, uncomment if needed

# Visualization
matplotlib>=3.8.0
opencv-python>=4.8.0
seaborn>=0.13.0

# Config & Utils
PyYAML>=6.0.1
tqdm>=4.66.0
numpy>=1.24.0
scipy>=1.11.0
